refactor(whatsrust): extract message caches
